requirements.txt
Flask
androguard
Pillow
imagehash
rapidfuzz